+++ packages/headless/src/api/service/insight/get-interface/get-interface-config-request.ts
@@ -1,0 +1,13 @@
+import {BaseParam} from '../../../platform-service-params';
+import {baseInsightRequest, InsightIdParam} from '../insight-params';
+
+export type GetInsightInterfaceConfigRequest = BaseParam & InsightIdParam;
+
+export const buildGetInsightInterfaceConfigRequest = (
+  req: GetInsightInterfaceConfigRequest
+) => {
+  return {
+    ...baseInsightRequest(req, 'GET', 'application/json', '/interface'),
+    requestParams: {},
+  };
+};